requirements.txt
aiogram==3.5.0
python-dotenv==1.0.1
aiosqlite==3.1.1
telethon==1.35.1
cryptg==0.4.1
